+++ packages/client/app/(withLayout)/account/page.tsx
@@ -1,5 +1,4 @@
 import { AccountPage } from '@widgets/Account'
-import React from 'react'
 
 function Account() {
    return <AccountPage />
