+++ packages/client/src/clientPages/accountQuestion/index.ts
@@ -1,0 +1,1 @@
+export { ClientAccountQuestion } from './ui/ClientAccountQuestionPage'
