+++ packages/client/src/clientPages/accountScrap/index.ts
@@ -1,0 +1,1 @@
+export { ClientAccountBookmark } from './ui/ClientAccountScrapPage'
