+++ packages/client/src/clientPages/accountScrap/ui/index.css.ts
@@ -1,5 +1,11 @@
 import { style } from '@vanilla-extract/css'
 
+export const Wrapper = style({
+   display: 'flex',
+   flexDirection: 'column',
+   padding: '0 16px 16px 16px',
+})
+
 export const ListWrapper = style({
    display: 'flex',
    flexDirection: 'column',
